+++ internal/producer/src/producer/producer_test.go
@@ -7,23 +7,20 @@ import (
 	"crypto/rand"
 	"database/sql"
 	"encoding/binary"
-	"encoding/hex"
 	"fmt"
 	"net"
 	"os"
 	"reflect"
 	"testing"
 
+	"github.com/SIGBlockchain/project_aurum/internal/accountstable"
+	"github.com/SIGBlockchain/project_aurum/internal/blockchain"
 	"github.com/SIGBlockchain/project_aurum/internal/client/src/client"
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
 	"github.com/SIGBlockchain/project_aurum/internal/genesis"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accountinfo"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accountstable"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/block"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/blockchain"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/validation"
 	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 	"github.com/SIGBlockchain/project_aurum/internal/sqlstatements"
 )
@@ -37,59 +34,6 @@ func TestCheckConnectivity(t *testing.T) {
 		t.Errorf("Internet connection check failed.")
 	}
 }
-
-// Tests a single connection
-// func TestAcceptConnections(t *testing.T) {
-// 	ln, _ := net.Listen("tcp", "localhost:10000")
-// 	var buffer bytes.Buffer
-// 	bp := BlockProducer{
-// 		Server:        ln,
-// 		NewConnection: make(chan net.Conn, 128),
-// 		Logger:        log.New(&buffer, "LOG:", log.Ldate),
-// 	}
-// 	go bp.AcceptConnections()
-// 	conn, err := net.Dial("tcp", ":10000")
-// 	if err != nil {
-// 		t.Errorf("Failed to connect to server")
-// 	}
-// 	contentsOfChannel := <-bp.NewConnection
-// 	actual := contentsOfChannel.RemoteAddr().String()
-// 	expected := conn.LocalAddr().String()
-// 	if actual != expected {
-// 		t.Errorf("Failed to store connection")
-// 	}
-// 	conn.Close()
-// 	ln.Close()
-// }
-
-// Sends a message to the listener and checks to see if it echoes back
-// func TestHandler(t *testing.T) {
-// 	ln, _ := net.Listen("tcp", "localhost:10000")
-// 	var buffer bytes.Buffer
-// 	bp := BlockProducer{
-// 		Server:        ln,
-// 		NewConnection: make(chan net.Conn, 128),
-// 		Logger:        log.New(&buffer, "LOG:", log.Ldate),
-// 	}
-// 	go bp.AcceptConnections()
-// 	go bp.WorkLoop()
-// 	conn, err := net.Dial("tcp", ":10000")
-// 	if err != nil {
-// 		t.Errorf("Failed to connect to server")
-// 	}
-// 	expected := []byte("This is a test.")
-// 	conn.Write(expected)
-// 	actual := make([]byte, len(expected))
-// 	_, readErr := conn.Read(actual)
-// 	if readErr != nil {
-// 		t.Errorf("Failed to read from socket.")
-// 	}
-// 	if bytes.Equal(expected, actual) == false {
-// 		t.Errorf("Message mismatch")
-// 	}
-// 	conn.Close()
-// 	ln.Close()
-// }
 
 func TestRunServer(t *testing.T) {
 	senderPrivateKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
@@ -447,504 +391,6 @@ func TestData_Deserialize(t *testing.T) {
 			}
 			if !reflect.DeepEqual(tt.d, someData) {
 				t.Errorf("Deserialized Data struct failed to match")
-			}
-		})
-	}
-}
-
-// func TestBringOnTheGenesis(t *testing.T) {
-// 	var pkhashes [][]byte
-// 	var datum []contracts.Contract
-// 	for i := 0; i < 100; i++ {
-// 		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-// 		someKeyPKHash := hashing.New(publickey.Encode(&someKey.PublicKey))
-// 		pkhashes = append(pkhashes, someKeyPKHash)
-// 		someAirdropContract, _ := contracts.New(1, nil, someKeyPKHash, 10, 0)
-// 		datum = append(datum, *someAirdropContract)
-// 	}
-// 	genny, _ := block.New(1, 0, make([]byte, 32), datum)
-// 	type args struct {
-// 		genesisPublicKeyHashes [][]byte
-// 		initialAurumSupply     uint64
-// 	}
-// 	tests := []struct {
-// 		name    string
-// 		args    args
-// 		want    block.Block
-// 		wantErr bool
-// 	}{
-// 		{
-// 			args: args{
-// 				genesisPublicKeyHashes: pkhashes,
-// 				initialAurumSupply:     1000,
-// 			},
-// 			want:    genny,
-// 			wantErr: false,
-// 		},
-// 	}
-// 	for _, tt := range tests {
-// 		t.Run(tt.name, func(t *testing.T) {
-// 			got, err := BringOnTheGenesis(tt.args.genesisPublicKeyHashes, tt.args.initialAurumSupply)
-// 			if (err != nil) != tt.wantErr {
-// 				t.Errorf("BringOnTheGenesis() error = %v, wantErr %v", err, tt.wantErr)
-// 				return
-// 			}
-// 			if !reflect.DeepEqual(got.Version, tt.want.Version) ||
-// 				!reflect.DeepEqual(got.Height, tt.want.Height) ||
-// 				!reflect.DeepEqual(got.PreviousHash, tt.want.PreviousHash) ||
-// 				!reflect.DeepEqual(got.MerkleRootHash, tt.want.MerkleRootHash) ||
-// 				!reflect.DeepEqual(got.DataLen, tt.want.DataLen) ||
-// 				!reflect.DeepEqual(got.Data, tt.want.Data) {
-// 				t.Errorf("BringOnTheGenesis() = %v, want %v", got, tt.want)
-// 			}
-// 			for i := range got.Data {
-// 				deserializedData := contracts.Contract{}
-// 				err := deserializedData.Deserialize(got.Data[i])
-// 				if err != nil {
-// 					t.Errorf("failed to deserialize data")
-// 				}
-// 				deserializedContract := &contracts.Contract{}
-// 				serializedDataBdy, _ := deserializedData.Serialize()
-// 				deserializedContract.Deserialize(serializedDataBdy)
-// 				if deserializedContract.Value != 10 {
-// 					t.Errorf("failed to distribute aurum properly")
-// 				}
-// 			}
-// 		})
-// 	}
-// }
-
-// func TestAirdrop(t *testing.T) {
-// 	var pkhashes [][]byte
-// 	for i := 0; i < 100; i++ {
-// 		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-// 		someKeyPKHash := hashing.New(publickey.Encode(&someKey.PublicKey))
-// 		pkhashes = append(pkhashes, someKeyPKHash)
-// 	}
-// 	genny, _ := BringOnTheGenesis(pkhashes, 1000)
-// 	type args struct {
-// 		blockchain   string
-// 		metadata     string
-// 		accounts     string
-// 		genesisBlock block.Block
-// 	}
-// 	tests := []struct {
-// 		name    string
-// 		args    args
-// 		wantErr bool
-// 	}{
-// 		{
-// 			args: args{
-// 				blockchain:   "blockchain.dat",
-// 				metadata:     constants.MetadataTable,
-// 				accounts:     constants.AccountsTable,
-// 				genesisBlock: genny,
-// 			},
-// 		},
-// 	}
-// 	for _, tt := range tests {
-// 		defer func() {
-// 			os.Remove(tt.args.metadata)
-// 			os.Remove(tt.args.blockchain)
-// 			os.Remove(tt.args.accounts)
-// 		}()
-// 		t.Run(tt.name, func(t *testing.T) {
-// 			if err := Airdrop(tt.args.blockchain, tt.args.metadata, tt.args.accounts, tt.args.genesisBlock); (err != nil) != tt.wantErr {
-// 				t.Errorf("Airdrop() error = %v, wantErr %v", err, tt.wantErr)
-// 			}
-// 			fileGenny, err := ioutil.ReadFile(tt.args.blockchain)
-// 			if err != nil {
-// 				t.Errorf("Failed to open file" + err.Error())
-// 			}
-// 			serializedGenny := genny.Serialize()
-// 			if !bytes.Equal(fileGenny[4:], serializedGenny) {
-// 				t.Errorf("Genesis block does not match file block")
-// 			}
-
-// 			db, err := sql.Open("sqlite3", tt.args.accounts)
-// 			if err != nil {
-// 				t.Errorf("Failed to open accounts table: " + err.Error())
-// 			}
-// 			defer db.Close()
-
-// 			rows, err := db.Query(sqlstatements.GET_PUB_KEY_HASH_BALANCE_NONCE_FROM_ACCOUNT_BALANCES)
-// 			if err != nil {
-// 				t.Errorf("failed to create rows for queries")
-// 			}
-// 			defer rows.Close()
-
-// 			var pkhCount int
-// 			var pkhStr string
-// 			var balance int
-// 			var nonce int
-// 			for rows.Next() {
-// 				rows.Scan(&pkhStr, &balance, &nonce)
-// 				pkhash, _ := hex.DecodeString(pkhStr)
-// 				if !bytes.Equal(pkhash, pkhashes[pkhCount]) {
-// 					t.Errorf("hashes don't match: %v != %v\n", pkhash, pkhashes[pkhCount])
-// 				}
-// 				if balance != 10 {
-// 					t.Errorf("balance does not match: %v != %v\n", balance, 10)
-// 				}
-// 				if nonce != 0 {
-// 					t.Errorf("nonce does not match: %v != %v\n", nonce, 0)
-// 				}
-// 				pkhCount++
-// 			}
-// 		})
-// 	}
-// }
-
-// func TestReadGenesisHashes(t *testing.T) {
-// 	GenerateGenesisHashFile(50)
-
-// 	tests := []struct {
-// 		name    string
-// 		want    [][]byte
-// 		wantErr bool
-// 	}{
-// 		{
-// 			wantErr: false,
-// 		},
-// 	}
-// 	for _, tt := range tests {
-// 		t.Run(tt.name, func(t *testing.T) {
-// 			got, err := ReadGenesisHashes()
-// 			if (err != nil) != tt.wantErr {
-// 				t.Errorf("ReadGenesisHashes() error = %v, wantErr %v", err, tt.wantErr)
-// 				return
-// 			}
-// 			if len(got) != 50 {
-// 				t.Errorf("wrong count on number of hashes")
-// 			}
-// 		})
-// 	}
-// }
-
-func TestRecoverBlockchainMetadata(t *testing.T) {
-	var ljr = "blockchain.dat"
-	var meta = constants.MetadataTable
-	var accts = constants.AccountsTable
-
-	if file, err := os.Create(ljr); err != nil {
-		t.Errorf("Failed to create file.")
-	} else {
-		file.Close()
-	}
-	defer func() {
-		if err := os.Remove(ljr); err != nil {
-			t.Errorf("failed to remove blockchain file")
-		}
-	}()
-
-	if conn, err := sql.Open("sqlite3", meta); err != nil {
-		t.Errorf("failed to create metadata file")
-	} else {
-		statement, _ := conn.Prepare(sqlstatements.CREATE_METADATA_TABLE)
-		statement.Exec()
-		conn.Close()
-	}
-	defer func() {
-		if err := os.Remove(meta); err != nil {
-			t.Errorf("failed to remove metadata file")
-		}
-	}()
-
-	if conn, err := sql.Open("sqlite3", accts); err != nil {
-		t.Errorf("failed to create accounts file")
-	} else {
-		statement, _ := conn.Prepare(sqlstatements.CREATE_ACCOUNT_BALANCES_TABLE)
-		statement.Exec()
-		conn.Close()
-	}
-
-	defer func() {
-		if err := os.Remove(accts); err != nil {
-			t.Errorf("failed to remove accounts file" + err.Error())
-		}
-	}()
-
-	var pkhashes [][]byte
-	for i := 0; i < 100; i++ {
-		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-		someKeyPKHash := hashing.New(publickey.Encode(&someKey.PublicKey))
-		pkhashes = append(pkhashes, someKeyPKHash)
-	}
-	genny, _ := genesis.BringOnTheGenesis(pkhashes, 1000)
-	if err := genesis.Airdrop(ljr, meta, constants.AccountsTable, genny); err != nil {
-		t.Errorf("airdrop failed")
-	}
-
-	type args struct {
-		ledgerFilename      string
-		metadataFilename    string
-		accountBalanceTable string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			args: args{
-				ledgerFilename:      ljr,
-				metadataFilename:    meta,
-				accountBalanceTable: accts,
-			},
-		},
-	}
-	var blockchainHeightIdx = 0
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := RecoverBlockchainMetadata(tt.args.ledgerFilename, tt.args.metadataFilename, tt.args.accountBalanceTable); (err != nil) != tt.wantErr {
-				t.Errorf("RecoverBlockchainMetadata() error = %v, wantErr %v", err, tt.wantErr)
-			}
-			blockchainGenesisBlockSerialized, err := blockchain.GetBlockByHeight(blockchainHeightIdx, ljr, meta)
-			if err != nil {
-				t.Errorf("failed to get genesis block")
-			}
-			blockchainGenesisBlockDeserialized := block.Deserialize(blockchainGenesisBlockSerialized)
-			if !reflect.DeepEqual(blockchainGenesisBlockDeserialized, genny) {
-				t.Errorf("genesis blocks do not match")
-			}
-			dbc, _ := sql.Open("sqlite3", accts)
-			defer func() { // not sure if this defer will happen before the others, is it stack based?
-				if err := dbc.Close(); err != nil {
-					t.Errorf("Failed to close database: %s", err)
-				}
-			}()
-			for _, hsh := range pkhashes {
-				var pkhash string
-				var balance uint64
-				var nonce uint64
-				foundKey := false
-				rows, err := dbc.Query(sqlstatements.GET_PUB_KEY_HASH_BALANCE_NONCE_FROM_ACCOUNT_BALANCES)
-				if err != nil {
-					t.Errorf("Failed to acquire rows from table")
-				}
-				for rows.Next() {
-					err = rows.Scan(&pkhash, &balance, &nonce)
-					if err != nil {
-						t.Errorf("failed to scan rows: %s", err)
-					}
-					decodedPkhash, err := hex.DecodeString(pkhash)
-					if err != nil {
-						t.Errorf("failed to decode public key hash")
-					}
-					if bytes.Equal(hsh, decodedPkhash) {
-						foundKey = true
-						if balance != 10 {
-							t.Errorf("wrong balance on key: %v", hsh)
-						}
-						if nonce != 0 {
-							t.Errorf("wrong nonce on key: %v", hsh)
-						}
-					}
-				}
-				if !foundKey {
-					t.Errorf("Key not found in table: %v", hsh)
-				}
-			}
-		})
-		blockchainHeightIdx++
-	}
-}
-
-func TestRecoverBlockchainMetadata_TwoBlocks(t *testing.T) {
-	var ljr = "blockchain.dat"
-	var meta = constants.MetadataTable
-	var accts = constants.AccountsTable
-
-	// Create ledger file and the two tables
-	file, err := os.Create(ljr)
-	if err != nil {
-		t.Errorf("Failed to create file.")
-	} else {
-		file.Close()
-	}
-	metaDB, err := sql.Open("sqlite3", meta)
-	if err != nil {
-		t.Errorf("failed to create metadata file")
-	} else {
-		metaDB.Exec(sqlstatements.CREATE_METADATA_TABLE)
-		metaDB.Close()
-	}
-	acctsDB, err := sql.Open("sqlite3", accts)
-	if err != nil {
-		t.Errorf("failed to create accounts file")
-	} else {
-		acctsDB.Exec(sqlstatements.CREATE_ACCOUNT_BALANCES_TABLE)
-	}
-
-	defer func() {
-		if err := os.Remove(ljr); err != nil {
-			t.Errorf("failed to remove blockchain file")
-		}
-		if err := os.Remove(meta); err != nil {
-			t.Errorf("failed to remove metadata file")
-		}
-		if err := os.Remove(accts); err != nil {
-			t.Errorf("failed to remove accounts file")
-		}
-	}()
-
-	var pkhashes [][]byte
-	somePVKeys := make([]*ecdsa.PrivateKey, 3) // Grab 3 private keys for creating contracts
-	for i := 0; i < 100; i++ {
-		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-		someKeyPKHash := hashing.New(publickey.Encode(&someKey.PublicKey))
-		pkhashes = append(pkhashes, someKeyPKHash)
-		if i < 3 {
-			somePVKeys[i] = someKey
-		}
-	}
-	genesisBlk, _ := genesis.BringOnTheGenesis(pkhashes, 1000)
-	if err := genesis.Airdrop(ljr, meta, accts, genesisBlk); err != nil {
-		t.Errorf("airdrop failed")
-	}
-	// Insert pkhashes into account table for contract validation
-	for i := 0; i < 100; i++ {
-		err := accountstable.InsertAccountIntoAccountBalanceTable(acctsDB, pkhashes[i], 10)
-		if err != nil {
-			t.Errorf("Failed to insert pkhash (%v) into account table: %s", pkhashes[i], err.Error())
-		}
-	}
-
-	// Create 3 contracts
-	contrcts := make([]contracts.Contract, 3)
-
-	// Contract 1
-	recipPKHash := hashing.New(publickey.Encode(&(somePVKeys[1].PublicKey)))
-	contract1, _ := contracts.New(1, somePVKeys[0], recipPKHash, 5, 1) // pkh1 to pkh2
-	contract1.Sign(somePVKeys[0])
-	err = validation.ValidateContract(contract1)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	senderPKHash := hashing.New(publickey.Encode(&(somePVKeys[0].PublicKey)))
-	accountstable.ExchangeBetweenAccountsUpdateAccountBalanceTable(acctsDB, senderPKHash, recipPKHash, 5) // update accts table for further contracts
-
-	// Contract 2
-	recipPKHash = hashing.New(publickey.Encode(&(somePVKeys[2].PublicKey)))
-	contract2, _ := contracts.New(1, somePVKeys[1], recipPKHash, 7, 2) // pkh2 to pkh3
-	contract2.Sign(somePVKeys[1])
-	err = validation.ValidateContract(contract2)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	senderPKHash = hashing.New(publickey.Encode(&somePVKeys[1].PublicKey))
-	accountstable.ExchangeBetweenAccountsUpdateAccountBalanceTable(acctsDB, senderPKHash, recipPKHash, 7) // update accts table for further contracts
-
-	// Contract 3
-	recipPKHash = hashing.New(publickey.Encode(&(somePVKeys[1].PublicKey)))
-	contract3, _ := contracts.New(1, somePVKeys[2], recipPKHash, 5, 2) // pkh3 to pkh2
-	contract3.Sign(somePVKeys[2])
-	err = validation.ValidateContract(contract3)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	senderPKHash = hashing.New(publickey.Encode(&somePVKeys[2].PublicKey))
-	accountstable.ExchangeBetweenAccountsUpdateAccountBalanceTable(acctsDB, senderPKHash, recipPKHash, 5) // update accts table
-	acctsDB.Close()
-
-	contrcts[0] = *contract1
-	contrcts[1] = *contract2
-	contrcts[2] = *contract3
-
-	firstBlock, err := block.New(1, 1, block.HashBlock(genesisBlk), contrcts)
-	if err != nil {
-		t.Errorf("failed to create first block")
-	}
-	err = blockchain.AddBlock(firstBlock, ljr, meta)
-	if err != nil {
-		t.Errorf("failed to add first block")
-	}
-
-	os.Remove(meta)
-	os.Remove(accts)
-	type args struct {
-		ledgerFilename      string
-		metadataFilename    string
-		accountBalanceTable string
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			args: args{
-				ledgerFilename:      ljr,
-				metadataFilename:    meta,
-				accountBalanceTable: accts,
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := RecoverBlockchainMetadata(tt.args.ledgerFilename, tt.args.metadataFilename, tt.args.accountBalanceTable); err != nil {
-				t.Errorf("RecoverBlockchainMetadata() error = %v", err)
-			}
-			firstBlockSerialized, err := blockchain.GetBlockByHeight(1, ljr, meta)
-			if err != nil {
-				t.Errorf("failed to get firstBlock block")
-			}
-			firstBlockDeserialized := block.Deserialize(firstBlockSerialized)
-			if !reflect.DeepEqual(firstBlockDeserialized, firstBlock) {
-				t.Errorf("first blocks do not match")
-			}
-
-			dbc, _ := sql.Open("sqlite3", accts)
-			defer func() { // not sure if this defer will happen before the others, is it stack based?
-				if err := dbc.Close(); err != nil {
-					t.Errorf("Failed to close database: %s", err)
-				}
-			}()
-
-			for i, key := range somePVKeys {
-				someKeyPKhsh := hashing.New(publickey.Encode(&key.PublicKey))
-				var balance uint64
-				var nonce uint64
-				queryStr := fmt.Sprintf(sqlstatements.GET_BALANCE_NONCE_FROM_ACCOUNT_BALANCES_BY_PUB_KEY_HASH, hex.EncodeToString(someKeyPKhsh))
-				row, err := dbc.Query(queryStr)
-				if err != nil {
-					t.Errorf("Failed to acquire row from table")
-				}
-				if row.Next() {
-					err = row.Scan(&balance, &nonce)
-					if err != nil {
-						t.Errorf("failed to scan row: %s", err)
-					}
-					switch i {
-					case 0: // first contract
-						if balance != 5 { // 10 - 5
-							t.Errorf("wrong balance (%v) on key: %v", balance, someKeyPKhsh)
-						}
-						if nonce != 1 {
-							t.Errorf("wrong nonce (%v) on key: %v", nonce, someKeyPKhsh)
-						}
-						break
-					case 1: // second contract
-						if balance != 13 { // 10 + 5 - 7 + 5
-							t.Errorf("wrong balance (%v) on key: %v", balance, someKeyPKhsh)
-						}
-						if nonce != 3 {
-							t.Errorf("wrong nonce (%v) on key: %v", nonce, someKeyPKhsh)
-						}
-						break
-					default: // third contract
-						if balance != 12 { // 10 + 7 - 5
-							t.Errorf("wrong balance (%v) on key: %v", balance, someKeyPKhsh)
-						}
-						if nonce != 2 {
-							t.Errorf("wrong nonce (%v) on key: %v", nonce, someKeyPKhsh)
-						}
-						break
-					}
-				} else {
-					t.Errorf("Key not found in table: %v", someKeyPKhsh)
-				}
-				row.Close()
 			}
 		})
 	}
